requirements.txt
pypinyin~=0.41.0
pandas~=1.1.5
numpy~=1.19.5
paho-mqtt~=1.5.1